Activate new Ingress Helm-Chart
After the OSH-I upversion, the Ingress Helm chart was
changed from the one from Openstack community to the
Ingress-nginx. A previous commit [1] added the new
Helm chart to the build environment and another [2]
created the manifest for this new Ingress chart. Now,
this review removes the old Ingress manifest and activates
the new Ingress Helm chart.

A patch was created to rename the Ingress Helm-chart so it
wouldn't have a conflict with the Ingress deployed by the platform.

All manifests and plugins of the old Ingress was deleted, and
all references (kustomization, dependencies, etc...) was changed
to the new Ingress chart name.

The new Ingress was placed as a dependency to the OSH-I debian package.
This dependencie was moved to the stx-openstack-helm-fluxcd debian
control, to facilitate the tarball building proccess. As it was, the
new Ingress needed to be declared in the tarball building call, with
this change this is not necessary anymore.

Finally the name for the ingressClass name in all Helm charts' Ingress
was changed for the new name defined by the Ingress patch.

[1] - https://review.opendev.org/c/starlingx/openstack-armada-app/+/937396
[2] - https://review.opendev.org/c/starlingx/openstack-armada-app/+/937174

Test Plan:
PASS - Build all packages and STX-O tarball
PASS - STX-O upload and apply
PASS - New Ingress Helm chart deployed with new name
PASS - Ingress from all charts with new ingressclassname
PASS - Ingress comunicates with other Helm charts
PASS - Launch a VM

This is the last change in the realtion chain related to the upversion
to the Caracal version and the activation of the new Ingress chart. All
reviews must be merged together, otherwise the build and apply of STX-O
will break.

Story: 2011303
Task: 51611

Change-Id: If10a47a470b6e32ccb1090dbaa604fcbf6d1f82e
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>