Align Horizon image and overrides with 2024.1
This update modifies the Horizon image file to track the latest commit
of stable/2024.1, enabling the build of the stx-horizon Docker image.
It also updates the static overrides for MariaDB to upgrade it, as
Django 4.2 of Horizon requirements for 2024.1 mandate MariaDB 10.4 or
higher.

Additionally, this change includes pymemcache in the image and aligns
various components, including Horizon static overrides, to ensure
compatibility with pymemcache and the new version of Django. These
updates align with the stable/2024.1 branch of Horizon.

Test Plan:
- PASS: Successfully build the stx-horizon Docker image
- PASS: Apply the new tarball containing static override changes
- PASS: Override stx-horizon image tags
- PASS: Apply stx-openstack with Horizon and MariaDB image overrides

Story: 2011303
Task: 51502

Depends-on: https://review.opendev.org/c/starlingx/root/+/940525

Change-Id: I7be859eda41c0a22cb3b1cb6f0fe3180c7e2f492
Signed-off-by: Mateus Nascimento <mateus.soaresdonascimento@windriver.com>