Removed images from unsupported charts
stx-ironic image is currently failing to build on the f/caracal branch.

The charts aodh, barbican, ceilimoter and ironic are not currently
supported by STX-Openstack. Despite that, the stx-* images for this
charts are present on the list of images to be built.

This review removes the 4 images from the debian_stable_docker_images.inc
file. This way, unecessary images stops to being built

Test Plan:
PASS - Run build-stx-images.sh to build all images
PASS - No errors are given by the bash file
PASS - aodh, barbican, ceilometer and ironic images are not built

Closes-Bug: 2097783

Change-Id: If272117ce3983ec040cb508ca94ddf5b41954e3e
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>