Remove unused default Agents Daemonset
This change aims to remove the default daemonsets agents that are being
created but are not being used.
The _daemonset_overrides.tpl file, creates the daemonsets with
overrides and still creates the default daemonsets that end up not
being used.

The solution for this was by patching openstack-helm-infra removing a
section that aggregate the default daemonset to a global list.
In order to not remove the neutron-netns-cleanup-cron-default, an
override for this was created in the static overrides under the name
neutron-netns-cleanup-cron-openstack-control-plane.

Before the patch we had 25 daemonsets and 8 of them were not being used.
After the patch been applied we now have 17 daemonsets and all of them are being used.

Test Plan:

PASS: build-pkg -c -l openstack
PASS: build openstack tarball
PASS: Upload and Apply openstack tarball on standard system
PASS: *default* daemonsets are not deployed in openstack ns
PASS: New daemonset neutron-netns-cleanup-cron-openstack-control-plane
      deployed in openstack ns
PASS: Launch 2x Guest instances (centos-guest VMs)
PASS: Access instances through a remote console
PASS: Ping instances from remote console (all to all)
PASS: Ping instances from external server in the management network

Story: 2011304
Task: 51652

Change-Id: I0f83a8bf43ad4803820fdca22f77ceb182f30e9f
Signed-off-by: kgoncalv <kayo.goncalvesdacosta@windriver.com>
Co-authored-by: Ingo Almendros <ingo.almendrosgirao@windriver.com>