Upversion openstack-pkg-tools to caracal
As part of the STX-Openstack upversion, it was noticed that this
package creates a .deb necessary for images build.
All patches were updated in order to make the package buildable in the
new version.
With that upversion, future images updates will happen easily.

NOTE: this was initially committed on [1] but an error occurred on the
build of other packages so it was reverted until now, when that error
was fixed. The problem was a missed file on patch 0002

Test Plan:
PASS: Build the openstack-pkg-tools  package
PASS: Build all  packages
PASS: Build all stx docker images

Story: 2011303
Task: 51552

[1] https://review.opendev.org/c/starlingx/openstack-armada-app/+/939474

Change-Id: I2f53aadec423264404c6c004e477d798da4ba86f
Signed-off-by: Romulo Leite <romulo.leite@windriver.com>