Update docker image for OVS community image
After the upversion of OSH and OSH-I Helm charts to the caracal version, the image docker.io/starlingx/stx-ovs:master-debian-stable-latest is no longer compatible with the openvswitch deployment. This causes a CrashLoopBackOff in the OVS pods and STX-Openstack does not apply.

The focus here is to change the actual openvswitch to the community
version.

Test plan:

PASS: build-pkgs -c -a -l openstack
PASS: apply stx-openstack on a vdm
PASS: apply stx-openstack on a lab
PASS: create 2 vm's and ping each other separately

Closes-bug: #2096992

Change-Id: I2e82928f2083d2ebd5f3c7df71a781667f9260d7
Signed-off-by: kgoncalv <kayo.goncalvesdacosta@windriver.com>