Add helmrelease patch action for app update
It was found that during the update process, after the
helmrelease is uninstalled, the FluxCD would automatically
redeploy the helmrelease that was uninstalled because of
the reconcile function of FluxCD.

To prevent this, this review adds a step to patch the helmrelease
and inform FluxCD that the Helm release is not to be reconciled.

Also, it was changed the moment where the cleanup function is called.
Before, it was being called as soon as the application apply began.
Now it will begin after the process of downloading the docker images
is finished. This will dimish the downtime and if the download failed
the application would not be down at any time.

Test Plan:
PASS - Build python3-k8sapp-openstack package
PASS - Build STX-Openstack tarball
PASS - Run system application-update
PASS - Application updated to 25.03

Story: 2011262
Task: 51657

Change-Id: I8fa90c7167f856df47d10a9fb55fcea08cb18ab6
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>