Copy and restore mariadb data on recovery
ISSUE:
Accordingly to MariaDB's website [1], downgrading between major
releases is not supported. This may cause several issues when
we perform a recovery process of our openstack application, since
there is no guarantee that a helm chart rollback will work.

In order to prevent downgrade problems, the solution was to
completely remove the newer mariadb's helm chart prior to re-applying
the older version of the application [2]. This leads to another
problem: the loss of the data stored in mariadb's PVC when the
chart is removed.

SOLUTION:
Here, we devised a mechanism to copy/restore the PVC data when a
recovery process is triggered.
First, we create a Volume Snapshot [3] when the application update
starts. If the update fails, prior to removing mariadb's chart, we
restore the snapshot by deleting the current PVC and creating a new
PVC with the same name but pointing to the snapshot as the data source.
After the apply operation is done, the snapshot is deleted.

TEST PLAN:
PASS - Build the packages and tarballs
PASS - Apply an older version of the application; create a new user
with the 'openstack user create' command
PASS - Run an update to a newer version built with some modifications
to cause the update to fail (horizon chart with wrong overrides)
PASS - Check if the snapshot was created
PASS - When the recovery process starts, check if the snapshot was
applied
PASS - When the older mariadb's chart is applied, access the pod and
check if the user created is in the database.
PASS - Happy path - Perform a successful update and check if the
user is in the database and if the snapshot was deleted
PASS - Test in a Duplex environment

REFERENCES:
[1] https://mariadb.com/kb/en/downgrading-between-major-versions-of-mariadb/
[2] https://review.opendev.org/c/starlingx/openstack-armada-app/+/944284
[3] https://kubernetes.io/docs/concepts/storage/volume-snapshots/

Story: 2011262
Task: 51826

Co-authored-by: Daniel Caires <DanielMarques.Caires@windriver.com>

Change-Id: I688c5894535b1850a3444b3a9e7f18ec7008f4dd
Signed-off-by: Joao Fracarolli <joao.vicentinifracarolli@windriver.com>