Fix neutron deployment with https configured
After the OSH upversion to caracal, a new pod was being deployed
in the controller for Neutron. The rpc server was added as part
of the Neutron Helm chart deployment.

The problem was that the patches 0007 and 0008 from OSH, was
removing some information related to TLS and, because this is a
new template, the rpc-server was not included in this cleanup.

This review adds the rpc-server to the list of templates that
are being cleaned by patches 0007 and 0008.

Test Plan:
PASS - Build OSH and STX-O tarball
PASS - Deploy STX-O in a system with https enabled
PASS - Neutron rpc-server pod is running
PASS - Create Networks and launch a VM

Closes-Bug: #2103801

Change-Id: I84140bbd957ca07a0a53cde2bf58e0ededdb914d
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>