Add tolerations to mariadb-controller pod
Recently a new pod was added to the MariaDB Helm chart[1].
This pod works as an auxiliary to the deployment of the
MariaDB servers.

It was found that if the Controller node had the taint
"node-role.kubernetes.io/control-plane:NoSchedule", the
mariadb-controller pod would be stuck in pending state.

This review aims to update the patch that was added creating
this new pod, where it adds the tolerations section to the
pod definition.

[1] - https://review.opendev.org/c/starlingx/openstack-armada-app/+/944245

Test Plan:
PASS - Build OSH and STX-O tarball
PASS - Apply STX-O in a standard system
PASS - mariadb-controller pod is running
PASS - MariaDB Helmrelease succesfully deployed
PASS - STX-O applied

Closes-Bug: #2104035

Change-Id: I05b8dfbaeebd247cf72340881428aece78197df5
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>