Removes unused docker images
By upversioning openstack-helm [1], openstack-helm-infra [2] and
integrating the new ingress-nginx chart [3], new docker images were
introduced to stx-openstack. Even though some of these images are not
used by the app, their image values are not set to null and they are
still downloaded during the application apply.

This change removes the following images that are currently not required
for stx-openstack:
- bitnami/kubectl
- ingress-nginx/kube-webhook-certgen
- ingress-nginx/opentelemetry

[1]https://review.opendev.org/c/starlingx/openstack-armada-app/+/941982
[2]https://review.opendev.org/c/starlingx/openstack-armada-app/+/941981
[3]https://review.opendev.org/c/starlingx/openstack-armada-app/+/936194

Test Plan:
[PASS] build stx-openstack packages and tarball
[PASS] apply stx-openstack tarball to a virtual AIO-SX system
[PASS] check sysinv log and docker registries to ensure that images
       removed are no longer downloaded.

Closes-Bug: #2104057

Change-Id: I8c0ff659c0d9ab4bba2117c54a1ae1576d517479
Signed-off-by: Alex Figueiredo <alex.fernandesfigueiredo@windriver.com>