Increase number of workers in glance-api pod
After the caracal upversion, some services started using
uWSGI in their pods. This change caused glance to be unable
to create big images. The default number of workers running in
the Glance pod is 1, because of that, when trying to create an
image that is too big, Glance will not be capable to handle the
request and it will trigger a graceful shutdown of the
pod causing the image create command to fail.

To revert this problem, it is necessary to increase the number
of workers and processes in the glance-api pod. The way that the
Glance Helm-chart works, only by changing the number of workers,
the number of processes will increse with the same number.

This review aims to set the number of workers to 2, this increase
in number of workers/process will increase the cpu consumption
by 1 milicore and RAM by ~100Mi, which showed to be enough to allow
glance to creates big images.

Test Plan:
PASS - Build stx-openstack-helm-fluxcd and STX-O tarball
PASS - Upload and apply STX-Openstack
PASS - Increase the size of filesystem partitions [1]
PASS - Create glance image using Windows 10 raw file(20GB)

[1] - https://docs.starlingx.io/storage/openstack/increase-the-size-of-raw-image-uploads-in-openstack-2170f501d2e7.html

Change-Id: I3b285eb394d07bd10c3a6d112d3e71047b6818b8
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>