Update Ingress to allow IPv6 deployment
Currently the ingress-nginx-openstack fails to apply
in a IPv6 system. Because the values.yaml file sets
the ipFamilies to IPv4, the STX-O apply fails with an
error on the deployment of the ingress pod.

To fix this issue, it is necessary to change the ipFamilyPolicy
to PreferDualStack as the kubernetes documentation [1] suggests.
The ipFamilies also needs to be ovewritten to remove the hard-coded
"IPv4" set in the values file. Kubernetes services are capable of
detecting the ipFamily, without the need of previously setting it
in the service template.

This review changes the ipFamilyPolicy to "PreferDualStack" and
removes the hard-code "IPv4" from the ipFamilies parameter.

[1] - https://kubernetes.io/docs/concepts/services-networking/dual-stack/#services

Story: 2011387
Task: 51869

Test Plan:
PASS - Build stx-openstack-helm-fluxcd and STX-O tarball
PASS - Upload STX-Openstack
PASS - Deploy Ingress Helm-chart in an IPv6 system
PASS - Apply STX-O in an IPv4 system
PASS - Run STX-Openstack Sanity in an IPv4 system

Currently is not possibly to do a full STX-O install
in an IPv6 system. The full STX-O apply in a IPv6 system
will be done by the next review in the relation chain.

Change-Id: I84ac034a00937f19e14d915c6ab3fd6b5ff03d33
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>