[ingress-nginx] adds imagePullSecrets config
After [1] the stx-openstack app stopped using the deprecated ingress
chart provided by the openstack-helm-infra and introduced a custom
ingress-nginx helm chart based on the upstream ingress-nginx chart [2].

Since [3] and [4] the version of the ingress-nginx/controller image used
by the stx-openstack app diverged from the version used by the stx
platform. As a result, the stx-openstack app isn't able to pull the
ingress-nginx/controller image and so the stx-openstack apply is
failing.

This issue is happening because the stx-openstack app isn't setting the
imagePullSecrets [5] required to authenticate and pull the
ingress-nginx/controller image from the local registry. As reference,
for the openstack-helm and openstack-helm-infra charts, the
imagePullSecrets is configured by the patch [6].

This change adds the "default-registry-key" to the list of
imagePullSecrets used by the ingress-nginx-openstack chart.

[1]https://review.opendev.org/c/starlingx/openstack-armada-app/+/941983
[2]https://github.com/kubernetes/ingress-nginx/tree/helm-chart-4.11.5/charts/ingress-nginx
[3]https://review.opendev.org/c/starlingx/nginx-ingress-controller-armada-app/+/940615
[4]https://review.opendev.org/c/starlingx/openstack-armada-app/+/945786
[5]https://github.com/kubernetes/ingress-nginx/blob/helm-chart-4.11.5/charts/ingress-nginx/values.yaml#L1179
[6]https://opendev.org/starlingx/openstack-armada-app/src/branch/master/openstack-helm-infra/debian/deb_folder/patches/0001-Add-imagePullSecrets-in-service-account.patch

Test Plan:
[PASS] Build stx-openstack packages and tarball
[PASS] Upload and apply stx-openstack app into a virtual AIO-SX system
[PASS] Upload and apply stx-openstack app into a virtual AIO-DX system

Closes-Bug: #2106541

Change-Id: I78b1f62f366de7da747fab7b1a940a2014bdb724
Signed-off-by: Alex Figueiredo <alex.fernandesfigueiredo@windriver.com>