Change uWSGI socket to allow IPv6 binding
After the Caracal update, some of the OSH Helm charts
begun to use uWSGI on their pods. From the Helm-charts
of this services the htto-socket for the uWSGI was being
configured with a default value of 0.0.0.0 to the IP binding.
Because of that, this services failed to be deployed in an
IPv6 system.

This change creates a new patch that changes the default socket
to allow binding with an IPv6 IP. The patch changes the value
for http-socket in the charts: Cinder, Glance, Heat, Neutron,
Nova and Barbican. Altough this last one is not currently supported
by STX-Openstack, there are plans to support it in the near future.

Story: 2011387
Task: 51870

Test Plan:
PASS - Build openstack-helm package and STX-O tarball
PASS - Upload and apply STX-O in an IPv6 system
PASS - Upload and apply STX-O in an IPv4 system
PASS - Run STX-Openstack Sanity in an IPv4 system

Change-Id: I5da36d9fe403c2984094677f4c7844112e91e2e8
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>