[rook-ceph] Constants, utils and image override
This change includes a set of constants and functions to be used by the
stx-openstack plugins and lifecycle in order to support the rook ceph
storage backend. It also adds a new override to the rook ceph client
image (ceph-config-helper) to enable the app to dinamically change the
client image and ensure client version compatibility with both the
baremetal and rook ceph deployments. Furthermore, the change adds unit
tests covering the new functions added for rook ceph.

The function is_rook_ceph_backend_available can be used by the app to
decide between the baremetal or rook ceph storage backends. It can also
be used to ensure that rook ceph app is successfully deployed and
configured.

The function is_rook_ceph_api_available can be used combined to the
function is_rook_ceph_backend_available to protect the access to the
rook ceph REST API, ensuring that the RESP API pods are running.

The function get_rook_ceph_uuid can be used to access the rook ceph API
and read the Rook Ceph secret UUID for storage backend configuration. It
relies on is_rook_ceph_api_available to protect the access to rook ceph
REST API.

The new override for the rook ceph client image was added in the clients
helm chart to make it simpler to the app/user to configure the ceph
client image. Based on this config, the app plugins can dynamically
change all the ceph-config-helper image config for all the charts using
it (glance, libvirt, nova, cinder, etc.). The default value and static
override was defined based on the ceph-config-helper image currently
used by rook ceph app [1].

[1]https://opendev.org/starlingx/app-rook-ceph/src/branch/master/stx-rook-ceph-helm/stx-rook-ceph-helm/fluxcd-manifests/rook-ceph-cluster/rook-ceph-cluster-static-overrides.yaml#L259

Test Plan:
[PASS] Build stx-openstack packages
[PASS] Unit tests

Since stx-openstack still doesn't support even uploading to the rook
ceph stx deployments, the full validation of the app and new functions
(including potential fixes) will occur in subsequent tasks.

Story: 2011388
Task: 51871

Change-Id: I00436f795c33f4f8fa9247a9dc10b053880cad9d
Signed-off-by: Alex Figueiredo <alex.fernandesfigueiredo@windriver.com>