[rook-ceph] storage backend activation
This change activates the stx-openstack app support for the rook ceph
storage backend. For this, the cinder, glance, libvirt and nova charts
were properly configured based on the type of ceph backend currently
available (baremetal or rook ceph). The rook ceph support has also
required the lifecycle_openstack to be changed to support copying the
ceph configmap either from the rook-ceph or kube-system namespace.
Furthermore, a new patch was added to the openstack-helm package to
ensure that nova storage-init job can create and set the size of the
related ceph pool for AIO-SX systems (size=1).

This change also improves the dynamic image override feature introduced
by [1], updating the images:tags overrides in a reusable
_update_image_tag_overrides method of the OpenstackBaseHelm base
class. Unit tests for this and the others methods in openstack.py are
planned for a near future.

[1]https://review.opendev.org/c/starlingx/openstack-armada-app/+/947531

Test Plan:
[PASS] build stx-openstack packages and tarball

Baremetal ceph deployment:
[PASS] upload and apply the stx-openstack tarball to a virtual AIO-SX
[PASS] check that ceph-config-helper image overrides are NOT applied
[PASS] create an cirros image
[PASS] create a volume from cirros image
[PASS] create a volume backup
[PASS] delete backup, volume and image

Rook ceph deployment:
[PASS] upload and apply the stx-openstack tarball to a AIO-SX
[PASS] check that ceph-config-helper image overrides are applied
[PASS] create an cirros image
[PASS] create a volume from cirros image
[PASS] create a volume backup
[PASS] delete backup, volume and image

Story: 2011388
Task: 52007

Depends-On: https://review.opendev.org/c/starlingx/root/+/947876

Change-Id: I461ea05f33ece776b2cd92f780882943b62f2168
Signed-off-by: Alex Figueiredo <alex.fernandesfigueiredo@windriver.com>