Add mechanism to store/load the vswitch type data
A mechanism to store and load the vswitch type data was implemented
to improve the performance of the functions that check the vswitch type
of the system by reading the compute node labels [1].
The data is stored by overriding Neutron's helm chart, adding a
'vswitch_type' field that contains the label read during the
application-apply process. The default value of this field is
set to the string 'none', so when this value is read, the plugin
will try to fetch the data by looping through all the available
openstack compute nodes/hosts in the system.
After that, the value is stored, so it can be reused without the
need to loop again.
Some unit tests had to be modified and new tests were developed.
This change also fixes an issue in the log that was always logging
'There are no openstack-enabled compute nodes'.
Finally, we refactored some methods to be included inside the
OpenstackBaseHelm class to allow for better reusability of the
vswitch-checking methods by any Helm Chart that inherits from
this class.

REFERENCES:
[1] https://review.opendev.org/c/starlingx/openstack-armada-app/+/948065

TEST PLAN:
PASS - Build python3-k8sapp-openstack package
PASS - Build stx-openstack-helm-fluxcd package
PASS - Build stx-openstack tarball
PASS - All existent unit tests passed
PASS - New unit tests passed
PASS - Try to apply stx-openstack in different situations on a VDM
environment:
 - Label 'openvswitch' only, apply succeeds,
chart value: 'openvswitch=enabled'
 - Labels 'openvswitch' + 'openvswitch-dpdk', apply fails,
chart value: 'none'
 - No labels, apply fails, chart value: 'none'

Story: 2011263
Task: 52261

Change-Id: I7e7a1e43137fe9a82d8fe5f15a218c3bf9c461dd
Signed-off-by: Joao Fracarolli <joao.vicentinifracarolli@windriver.com>