Add pre_apply_cleanup img to garbd overrides
Garbd static-overrides were missing the pre_apply_cleanup
image in its list. That caused the tag to not be updated
by the build script with the correct tags, causing an error
during the images download proccess.

This review adds the image, as it is in all of the others
static-overrides.

Test Plan:
PASS - build stx-openstack-helm-flxucd package and STX-O tarball
PASS - Verify that the image has correct tag
PASS - Download does not fail during app apply
PASS - Application applied

Closes-bug: 2112590

Change-Id: Ib354282368bc3e81063f9d2dbccd82b299f741ee
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>