Update the OVS-DPDK vswitch label and add support to label combinations
The 'openvswitch-dpdk=enabled' label was modified to 'dpdk=enabled' and
now it works together with the 'openvswitch=enabled' label to indicate
that we're using OVS-DPDK.
To do this, some logic was added to accept the inclusion of both labels.
The code now uses a list of allowed combinations and for each openstack
enabled host, it verifies if the vswitch labels are included in this
list. After that, it checks if there are label conflicts within the
host and between hosts.
This approach affects some other parts of the code, such as
the Lifecycle Semantic Check, so all the necessary changes were made,
including adapting some unit tests and creating new ones.

TEST PLAN:
PASS - Build python-k8sapp-openstack and stx-openstack-helm-fluxcd packages
PASS - Build STX-O tarball
PASS - All unit tests must pass
PASS - Apply STX-O in a virtual deployment environment with different configs:
 > OVS label only, select OVS, must apply
 > DPDK label only, select none, must fail
 > OVS+DPDK labels, select OVS+DPDK, must apply
 > No label, select none, must fail

Story: 2011263
Task: 52396

Change-Id: I4f9bc997358769d65a720dde9cd130f985ab61ea
Signed-off-by: Joao Fracarolli <joao.vicentinifracarolli@windriver.com>