Apply STX-Openstack with OVS-DPDK
Currently, STX-Openstack has no logic to block the
Openvswitch Helm chart to deploy if the OVS-DPDK is
enabled on the platform.

This review aims to allow the application framework to
block the Openvswitch Helm chart deployment if the
OVS-DPDK labels are enabled.

This review also adds the override in the openvswitch
Helm chart that deploys the dpdk container, for the
future implementation of conteinerized OVS-DPDK in
STX-Openstack

Test Plan:
PASS - Build plugins package and STX-O tarball
PASS - Apply STX-O in a AIO-SX and verify:
     > If only openvswitch label = Openvswitch chart deployed
     > If openvswitch and dpdk label = Don't deploy Openvswitch
     > If only dpdk label = Apply fail
     > If no label = Apply fail

Story: 2011263
Task: 51290

Change-Id: I7724cd81426984934d34bcc7b0101a8d14ab28c9
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>