Fixes application recovery and reconciliation
Currently, the app recovery process is crashing or not behaving as
expected depending on the moment where the update operation is aborted.
Most of the time, the recovery process crashes if the update operation
is aborted around 80%. Other times the recover process is marked as
completed by the StarlingX Application Framework (a.k.a AppFwk), but
some of the helmreleases are still not on the state "READY=True"
("kubectl get helmreleases -n openstack").

The following problems were identified:
- The AppFwk is not forcing the reconciliation of all the helmreleases
during the app recovery operation, as done for the update operation.
- The nginx_ports_control chart is considered as disabled by the AppFwk
during the recovery process since its helm plugin is not applying any
system overrides. Because of this, some helmreleases depending on the
nginx_ports_control (e.g., mariadb helmrelease) may not be reconciled
even when the recover process is considered as "completed" by the
AppFwk.

This change fixes those issues by:
- Forcing the recociliation of all the application helmreleases during
the recovery process using the same procedure applied by the AppFwk
during the update process.
- Applying a system override to the nginx_ports_control by using the
same config already applied as a static override. This way the
nginx_ports_control helmrelease is not erroneously considered as
disabled by the AppFwk.

Closes-Bug: 2113879

Test Plan (vbox AIO-SX):
[PASS] build application packages and tarball (25.09.0.tgz)
[PASS] Upload and apply the previous app version (25.03.0.tgz)
[PASS] Update from app 25.03.0 to 25.09.0
       (system application-update 25.09.0.tgz)
[PASS] Wait for the helm charts apply to achieve ~80%
       (system application-show stx-openstack) and abort the update
       operation (system application-abort stx-openstack)
[PASS] Wait for the recovery process to complete successfully
[PASS] All the helmreleases marked as "READY=True"
       (kubectl get helmreleases -n openstack)
[PASS] Update from 25.03.0 to 25.09.0
       (system application-update 25.09.0.tgz)
[PASS] Wait for the update process to complete successfully
       (system application-show stx-openstack)
[PASS] Wait for the helmreleases to be marked as "READY=True"
[PASS] Remove and delete the app

Change-Id: I771f3640fd258920e5c62a1ab03325a9e5724820
Signed-off-by: Alex Figueiredo <alex.fernandesfigueiredo@windriver.com>