Fix Openvswitch plugin overrides
A recently added override in the Openvswitch plugin
has a wrong structure, and this causes the chart to
fail during its deployment.

This change fixes the plugin adding the correct
structure to the override, based on the values file
from the Helm-chart.

Test Plan:
PASS - Build plugins package and STX-O tarball
PASS - Upload and apply STX Openstack
PASS - Verify the correct structure of the override
PASS - STX-O successfully applied

Closes-Bug: #2115997

Change-Id: I0a284c0f8c9535b1ec63246bfbfb3a654e156a87
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>