Provision host UUID to nova-compute
This change adds new overrides and an OSH patch to provision
nova-compute pods with their respective hosts' UUID. Before
this change, Nova was attributing a newly generated UUID for
each nova-compute pod, accordingly to the documentation [1].
To do this, we added a method in Nova's plugin that retrieves
all the hostnames and UUIDs and inject them via overrides
into Nova's Helm chart. The nova-compute-init script was
then modified to search the current host's UUID and save
it in the compute_id file in the state_path directory.

References:
[1] https://docs.openstack.org/nova/latest/admin/compute-node-identification.html

Test Plan:
 - PASS: build openstack-helm package
 - PASS: build python3-k8aspp-openstack package
 - PASS: generate application tarball
 - PASS: apply the application in a virtual deployment environment
 - PASS: check if the compute_id file is created with the correct UUID
 - PASS: check if the correct UUID is registered in the database
 - PASS: repeat the test in a lab with a couple of compute nodes
 - PASS: lock, reinstall and unlock one of the compute nodes and
   check if nova pods are running again, with the same UUID

Closes-Bug: 2117400

Change-Id: I4eb2e43b3465d1f92f9cc8ebf40018998fc054a8
Signed-off-by: Joao Fracarolli <joao.vicentinifracarolli@windriver.com>